Reverting to using i32 for the regid parameter in reg_read() and reg_write().
- added a test for mips emulation

+++ src/unicorn_const.rs
@@ -2,8 +2,8 @@
 
 pub const API_MAJOR: u32 = 0;
 pub const API_MINOR: u32 = 9;
-pub const SECOND_SCALE: u32 = 1000000;
-pub const MILISECOND_SCALE: u32 = 1000;
+pub const SECOND_SCALE: u64 = 1000000;
+pub const MILISECOND_SCALE: u64 = 1000;
 
 // Architecture type
 #[repr(C)]
